Refactor tests
- Add tests for the new Company and Person classes
- Introduce TestUtil classes for Customer and Account
