+++ src/test/java/com/danix/code/smell/example001/CustomerTest.java
@@ -1,8 +1,8 @@
 package com.danix.code.smell.example001;
 
-import org.junit.Before;
 import org.junit.Test;
 
+import static com.danix.code.smell.example001.CustomerTestUtils.getPersonWithAccount;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 
@@ -10,77 +10,6 @@ import static org.junit.Assert.assertThat;
  * @author danix
  */
 public class CustomerTest {
-
-    private CustomerFactory customerFactory;
-
-    @Before
-    public void setUp() {
-        customerFactory = new CustomerFactory();
-    }
-
-    @Test
-    public void testWithdrawPersonWithNormalAccount() throws Exception {
-        Account account = getAccountByTypeAndMoney(false, 34.0);
-        Customer customer = getPersonCustomer(account);
-        customer.withdraw(10, "EUR");
-        assertThat(account.getMoneyAmount(), is(24.0));
-    }
-
-    @Test
-    public void testWithdrawPersonWithNormalAccountAndOverdraft() throws Exception {
-        Account account = getAccountByTypeAndMoney(false, -10.0);
-        Customer customer = getPersonCustomer(account);
-        customer.withdraw(10, "EUR");
-        assertThat(account.getMoneyAmount(), is(-22.0));
-    }
-
-    @Test
-    public void testWithdrawPersonWithPremiumAccount() throws Exception {
-        Account account = getAccountByTypeAndMoney(true, 34.0);
-        Customer customer = getPersonCustomer(account);
-        customer.withdraw(10, "EUR");
-        assertThat(account.getMoneyAmount(), is(24.0));
-    }
-
-    @Test
-    public void testWithdrawPersonWithPremiumAccountAndOverdraft() throws Exception {
-        Account account = getAccountByTypeAndMoney(true, -10.0);
-        Customer customer = getPersonCustomer(account);
-        customer.withdraw(10, "EUR");
-        assertThat(account.getMoneyAmount(), is(-21.0));
-    }
-
-    @Test
-    public void testWithdrawCompanyWithNormalAccount() throws Exception {
-        Account account = getAccountByTypeAndMoney(false, 34);
-        Customer customer = getCompanyCustomer(account);
-        customer.withdraw(10, "EUR");
-        assertThat(account.getMoneyAmount(), is(24.0));
-    }
-
-    @Test
-    public void testWithdrawCompanyWithNormalAccountAndOverdraft() throws Exception {
-        Account account = getAccountByTypeAndMoney(false, -10);
-        Customer customer = getCompanyCustomer(account);
-        customer.withdraw(10, "EUR");
-        assertThat(account.getMoneyAmount(), is(-21.0));
-    }
-
-    @Test
-    public void testWithdrawCompanyWithPremiumAccount() throws Exception {
-        Account account = getAccountByTypeAndMoney(true, 34);
-        Customer customer = getCompanyCustomer(account);
-        customer.withdraw(10, "EUR");
-        assertThat(account.getMoneyAmount(), is(24.0));
-    }
-
-    @Test
-    public void testWithdrawCompanyWithPremiumAccountAndOverdraft() throws Exception {
-        Account account = getAccountByTypeAndMoney(true, -10);
-        Customer customer = getCompanyCustomer(account);
-        customer.withdraw(10, "EUR");
-        assertThat(account.getMoneyAmount(), is(-20.25));
-    }
 
     @Test
     public void testPrintCustomerDaysOverdrawn() throws Exception {
@@ -104,34 +33,5 @@ public class CustomerTest {
     public void testPrintCustomerAccountPremium() throws Exception {
         Customer customer = getPersonWithAccount(true);
         assertThat(customer.printCustomerAccount(), is("Account: IBAN: RO023INGB434321431241, Money: 34.0, Account type: premium"));
-    }
-
-    private Customer getPersonWithAccount(boolean premium) {
-        AccountType accountType = new AccountType(premium);
-        Account account = new Account(accountType, 9);
-        Customer customer = getPersonCustomer(account);
-        account.setIban("RO023INGB434321431241");
-        account.setMoney(Money.newEuro(34.0));
-        return customer;
-    }
-
-    private Account getAccountByTypeAndMoney(boolean premium, double money) {
-        AccountType accountType = new AccountType(premium);
-        Account account = new Account(accountType, 9);
-        account.setIban("RO023INGB434321431241");
-        account.setMoney(Money.newEuro(money));
-        return account;
-    }
-
-    private Customer getPersonCustomer(Account account) {
-        Customer customer = customerFactory.createPerson("danix", "dan", "dan@mail.com", account);
-        account.setCustomer(customer);
-        return customer;
-    }
-
-    private Customer getCompanyCustomer(Account account) {
-        Customer customer = customerFactory.createCompany("company", "company@mail.com", account, 0.50);
-        account.setCustomer(customer);
-        return customer;
     }
 }
